Extract TE-specific shader components from GLShader
