+++ te-app/src/main/java/titanicsend/pattern/glengine/GLShader.java
@@ -1,27 +1,35 @@
 package titanicsend.pattern.glengine;
 
-import static com.jogamp.opengl.GL.*;
+import static com.jogamp.opengl.GL.GL_ARRAY_BUFFER;
+import static com.jogamp.opengl.GL.GL_ELEMENT_ARRAY_BUFFER;
 
 import Jama.Matrix;
 import com.jogamp.common.nio.Buffers;
-import com.jogamp.opengl.*;
+import com.jogamp.opengl.GL;
 import com.jogamp.opengl.GL4;
+import com.jogamp.opengl.GLAutoDrawable;
 import com.jogamp.opengl.util.GLBuffers;
 import com.jogamp.opengl.util.texture.Texture;
 import heronarts.lx.LX;
-import heronarts.lx.model.LXModel;
-import heronarts.lx.parameter.LXParameter;
 import java.io.File;
-import java.nio.*;
-import java.util.*;
+import java.nio.ByteBuffer;
+import java.nio.FloatBuffer;
+import java.nio.IntBuffer;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
-import titanicsend.pattern.yoffa.shader_engine.*;
+import titanicsend.pattern.yoffa.shader_engine.FragmentShader;
+import titanicsend.pattern.yoffa.shader_engine.ShaderAttribute;
+import titanicsend.pattern.yoffa.shader_engine.ShaderProgram;
+import titanicsend.pattern.yoffa.shader_engine.Uniform;
+import titanicsend.pattern.yoffa.shader_engine.UniformType;
 
-/**
- * Everything to run a complete shader program with geometry, textures, uniforms, vertex and
- * fragment shaders.
- */
-public class GLShader {
+/** Shader program components that are currently common across shader classes. */
+public abstract class GLShader {
 
   /**
    * Callback interface to set any uniforms that have been modified since the last frame. Can be
@@ -42,111 +50,53 @@ public class GLShader {
 
   // Vertex index list for default geometry. Since we're drawing
   // with triangles, we need two to make our rectangle!
-  private static final int[] INDICES = {
+  protected static final int[] INDICES = {
     0, 1, 2,
     2, 0, 3
   };
 
-  // local copies of frequently used OpenGL objects
-  private final GLEngine glEngine;
+  // Local copies of frequently used OpenGL objects
+  protected final GLEngine glEngine;
   private GLAutoDrawable canvas = null;
-  private GL4 gl4 = null;
+  protected GL4 gl4 = null;
 
-  private final FragmentShader fragmentShader;
-  private final int width;
-  private final int height;
+  protected ShaderProgram shaderProgram;
+  protected final FragmentShader fragmentShader;
+  protected final int width;
+  protected final int height;
 
-  private ShaderProgram shaderProgram;
-
-  // geometry buffers
+  // Geometry buffers
   private final FloatBuffer vertexBuffer;
   private final IntBuffer indexBuffer;
-  private final int[] vaohandles = new int[1];
+  private final int[] vaoHandles = new int[1];
   private final int[] geometryBufferHandles = new int[2];
 
-  // texture buffers
-  private class TextureInfo {
-    public String name;
-    public int channel;
-    public int textureUnit;
-    public int uniformLocation;
-  }
-
-  private boolean useMappedBuffer = false;
-
-  private final ArrayList<TextureInfo> textures = new ArrayList<>();
-  private final ByteBuffer backBuffer;
-  private final int[] backbufferHandle = new int[1];
-
-  // support for optional texture mapped (as opposed to the new linear format) buffer
-  // containing the last rendered frame.
-  private ByteBuffer mappedBuffer = null;
-  private final int[] mappedBufferHandle = new int[1];
-  private int mappedBufferUnit = -1;
-  private int mappedBufferWidth = 640;
-  private int mappedBufferHeight = 640;
-
-  // the GL texture unit to which the current view model coordinate
-  // texture is bound.
-  private int modelCoordsTextureUnit = -1;
-  private boolean modelCoordsChanged = false;
-
+  // Uniforms
+  private final HashMap<String, Uniform> uniformMap = new HashMap<>();
   private final List<Uniform> mutableUniforms = new ArrayList<>();
   public final List<Uniform> uniforms = Collections.unmodifiableList(this.mutableUniforms);
-  // map of user created uniforms.
-  protected final HashMap<String, Uniform> uniformMap = new HashMap<>();
 
-  // maps uniform names to GL texture units
+  // Map of uniform names to GL texture units
   protected final HashMap<String, Integer> uniformTextureUnits = new HashMap<>();
 
   // Callbacks used by the pattern or effect that owns the shader to set any custom uniforms.
-  // Usually, these are the parameters associated with the TECommonControls.
   private final List<UniformSource> uniformSources = new ArrayList<>();
 
   private boolean initialized = false;
 
-  // get the active GL profile so the calling entity can work with
-  // GL textures and buffers if necessary.  (NDI support
-  // requires this, for example.)
-  public GLProfile getGLProfile() {
-    return gl4.getGLProfile();
-  }
-
-  // get the active GL4 object for this shader.
-  public GL4 getGL4() {
-    return gl4;
+  /** Helper to construct a FragmentShader inline */
+  protected static FragmentShader newFragmentShader(
+      String shaderFilename, String... textureFilenames) {
+    return new FragmentShader(
+        new File("resources/shaders/" + shaderFilename),
+        Arrays.stream(textureFilenames)
+            .map(x -> new File("resources/shaders/textures/" + x))
+            .collect(Collectors.toList()));
   }
 
   // Welcome to the Land of 1000 Constructors!
 
-  /**
-   * Create new OpenGL shader effect
-   *
-   * @param lx LX instance
-   * @param fragmentShader fragment shader object to use
-   * @param frameBuf native (GL compatible) ByteBuffer to store render results for use in shaders
-   *     that need to read the previous frame. If null, a buffer will be automatically allocated.
-   * @param uniformSource callback that will set uniforms on this shader
-   */
-  public GLShader(
-      LX lx, FragmentShader fragmentShader, ByteBuffer frameBuf, UniformSource uniformSource) {
-    this(lx, fragmentShader, frameBuf, List.of(uniformSource));
-  }
-
-  /**
-   * Create new OpenGL shader effect
-   *
-   * @param lx LX instance
-   * @param fragmentShader fragment shader object to use
-   * @param frameBuf native (GL compatible) ByteBuffer to store render results for use in shaders
-   *     that need to read the previous frame. If null, a buffer will be automatically allocated.
-   * @param uniformSources list of callbacks that will set uniforms on this shader
-   */
-  public GLShader(
-      LX lx,
-      FragmentShader fragmentShader,
-      ByteBuffer frameBuf,
-      List<UniformSource> uniformSources) {
+  public GLShader(LX lx, FragmentShader fragmentShader) {
 
     this.glEngine = (GLEngine) lx.engine.getChild(GLEngine.PATH);
     this.width = this.glEngine.getWidth();
@@ -158,8 +108,18 @@ public class GLShader {
     this.indexBuffer = Buffers.newDirectIntBuffer(INDICES.length);
     this.vertexBuffer.put(VERTICES);
     this.indexBuffer.put(INDICES);
-    // allocate default buffer for reading offscreen surface to cpu memory
-    this.backBuffer = frameBuf != null ? frameBuf : allocateBackBuffer();
+  }
+
+  /**
+   * Create new OpenGL shader effect
+   *
+   * @param lx LX instance
+   * @param fragmentShader fragment shader object to use that need to read the previous frame. If
+   *     null, a buffer will be automatically allocated.
+   * @param uniformSources list of callbacks that will set uniforms on this shader
+   */
+  public GLShader(LX lx, FragmentShader fragmentShader, List<UniformSource> uniformSources) {
+    this(lx, fragmentShader);
 
     for (UniformSource uniformSource : uniformSources) {
       if (uniformSource != null) {
@@ -169,14 +129,14 @@ public class GLShader {
   }
 
   /**
-   * Create new shader object with default backbuffer
+   * Create new OpenGL shader effect
    *
    * @param lx LX instance
-   * @param fragmentShader fragment shader object shader to use
+   * @param fragmentShader fragment shader object to use
    * @param uniformSource callback that will set uniforms on this shader
    */
   public GLShader(LX lx, FragmentShader fragmentShader, UniformSource uniformSource) {
-    this(lx, fragmentShader, null, uniformSource);
+    this(lx, fragmentShader, List.of(uniformSource));
   }
 
   /**
@@ -184,27 +144,6 @@ public class GLShader {
    *
    * @param lx LX instance
    * @param shaderFilename shader to use
-   * @param frameBuf native (GL compatible) ByteBuffer to store render results for use in shaders
-   *     that need to read the previous frame. If null, a buffer will be automatically allocated. *
-   * @param uniformSource callback that will set uniforms on this shader
-   * @param textureFilenames (optional) texture files to load
-   */
-  public GLShader(
-      LX lx,
-      String shaderFilename,
-      UniformSource uniformSource,
-      ByteBuffer frameBuf,
-      String... textureFilenames) {
-    this(lx, shaderFilename, List.of(uniformSource), frameBuf, textureFilenames);
-  }
-
-  /**
-   * Creates new shader object with additional texture support
-   *
-   * @param lx LX instance
-   * @param shaderFilename shader to use
-   * @param frameBuf native (GL compatible) ByteBuffer to store render results for use in shaders
-   *     that need to read the previous frame. If null, a buffer will be automatically allocated. *
    * @param uniformSources list of callbacks that will set uniforms on this shader
    * @param textureFilenames (optional) texture files to load
    */
@@ -212,17 +151,8 @@ public class GLShader {
       LX lx,
       String shaderFilename,
       List<UniformSource> uniformSources,
-      ByteBuffer frameBuf,
       String... textureFilenames) {
-    this(
-        lx,
-        new FragmentShader(
-            new File("resources/shaders/" + shaderFilename),
-            Arrays.stream(textureFilenames)
-                .map(x -> new File("resources/shaders/textures/" + x))
-                .collect(Collectors.toList())),
-        frameBuf,
-        uniformSources);
+    this(lx, newFragmentShader(shaderFilename, textureFilenames), uniformSources);
   }
 
   /**
@@ -235,23 +165,7 @@ public class GLShader {
    */
   public GLShader(
       LX lx, String shaderFilename, UniformSource uniformSource, String... textureFilenames) {
-    this(lx, shaderFilename, uniformSource, null, textureFilenames);
-  }
-
-  /**
-   * Creates new shader object with additional texture support
-   *
-   * @param lx LX instance
-   * @param shaderFilename shader to use
-   * @param uniformSources list of callbacks that will set uniforms on this shader
-   * @param textureFilenames (optional) texture files to load
-   */
-  public GLShader(
-      LX lx,
-      String shaderFilename,
-      List<UniformSource> uniformSources,
-      String... textureFilenames) {
-    this(lx, shaderFilename, uniformSources, null, textureFilenames);
+    this(lx, shaderFilename, List.of(uniformSource), textureFilenames);
   }
 
   // Buffers
@@ -280,23 +194,6 @@ public class GLShader {
 
   // Setup
 
-  /**
-   * Set the buffer to be used as a rectangular texture backbuffer for this shader. (As opposed to
-   * iBackbuffer, which is a linear list of colors corresponding to LX 3D model points and can't be
-   * used for algorithms that need the ability to access neighboring pixels.) NOTE: MUST BE CALLED
-   * BEFORE THE SHADER IS INITIALIZED. (i.e. in the pattern's constructor.) TODO - at present, this
-   * buffer is only used by shader effects. It should eventually TODO - be optional for shader
-   * patterns as well.
-   *
-   * @param buffer previously allocated ByteBuffer of sufficient size to hold the desired texture
-   */
-  public void setMappedBuffer(ByteBuffer buffer, int width, int height) {
-    this.mappedBufferWidth = width;
-    this.mappedBufferHeight = height;
-    this.mappedBuffer = buffer;
-    this.useMappedBuffer = true;
-  }
-
   protected GLShader addUniformSource(UniformSource uniformSource) {
     if (uniformSource == null) {
       throw new IllegalArgumentException("UniformSource cannot be null");
@@ -305,13 +202,9 @@ public class GLShader {
     return this;
   }
 
-  public List<ShaderConfiguration> getShaderConfig() {
-    return fragmentShader.getShaderConfig();
-  }
-
-  /** List of LX control parameters from the shader code */
-  public List<LXParameter> getParameters() {
-    return this.fragmentShader.parameters;
+  // get the active GL4 object for this shader.
+  public GL4 getGL4() {
+    return gl4;
   }
 
   // Initialization
@@ -333,7 +226,7 @@ public class GLShader {
   }
 
   /** Shader initialization that requires the OpenGL context. Call once. */
-  public void init() {
+  public final void init() {
     if (this.initialized) {
       throw new IllegalStateException("Shader already initialized");
     }
@@ -352,24 +245,21 @@ public class GLShader {
     this.canvas.getContext().makeCurrent();
     initShaderProgram();
     allocateShaderBuffers();
-    loadTextureFiles();
   }
 
   private void initShaderProgram() {
     this.shaderProgram = new ShaderProgram(this.gl4, this.fragmentShader.getShaderName());
   }
 
-  /**
-   * Called at pattern initialization time to allocate and configure GPU buffers that are common to
-   * all shaders.
-   */
-  private void allocateShaderBuffers() {
+  protected void allocateShaderBuffers() {
+    // Shader geometry is the same across all our shader classes, for now.
+
     // storage for geometry buffer handles
-    gl4.glGenVertexArrays(1, this.vaohandles, 0);
+    gl4.glGenVertexArrays(1, this.vaoHandles, 0);
     gl4.glGenBuffers(2, IntBuffer.wrap(geometryBufferHandles));
 
-    // Bind Vertex Array Object first
-    gl4.glBindVertexArray(this.vaohandles[0]);
+    // bind Vertex Array Object first
+    gl4.glBindVertexArray(this.vaoHandles[0]);
 
     // vertices
     vertexBuffer.rewind();
@@ -393,61 +283,8 @@ public class GLShader {
     gl4.glVertexAttribPointer(position, 3, GL4.GL_FLOAT, false, 0, 0);
     gl4.glEnableVertexAttribArray(position);
 
+    // finished with VAO
     gl4.glBindVertexArray(0);
-
-    // backbuffer texture object
-    gl4.glActiveTexture(GL_TEXTURE2);
-    gl4.glEnable(GL_TEXTURE_2D);
-    gl4.glGenTextures(1, backbufferHandle, 0);
-    gl4.glBindTexture(GL4.GL_TEXTURE_2D, backbufferHandle[0]);
-
-    gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-    gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
-    gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-    gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-
-    gl4.glBindTexture(GL_TEXTURE_2D, 0);
-
-    // create mapped buffer texture object, if needed
-    if (useMappedBuffer) {
-      this.mappedBufferUnit = glEngine.getNextTextureUnit();
-
-      gl4.glActiveTexture(GL_TEXTURE0 + mappedBufferUnit);
-      gl4.glEnable(GL_TEXTURE_2D);
-      gl4.glGenTextures(1, mappedBufferHandle, 0);
-      gl4.glBindTexture(GL4.GL_TEXTURE_2D, mappedBufferHandle[0]);
-
-      gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-      gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-      gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-      gl4.glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-
-      gl4.glBindTexture(GL_TEXTURE_2D, 0);
-    }
-
-    // assign shared uniform blocks to the shader's binding points
-    int perRunBlockIndex = gl4.glGetUniformBlockIndex(shaderProgram.id, "PerRunBlock");
-    gl4.glUniformBlockBinding(
-        shaderProgram.id, perRunBlockIndex, GLEngine.perRunUniformBlockBinding);
-
-    int perFrameBlockIndex = gl4.glGetUniformBlockIndex(shaderProgram.id, "PerFrameBlock");
-    gl4.glUniformBlockBinding(
-        shaderProgram.id, perFrameBlockIndex, GLEngine.perFrameUniformBlockBinding);
-  }
-
-  private void loadTextureFiles() {
-    for (Map.Entry<Integer, String> textureInput :
-        this.fragmentShader.getChannelToTexture().entrySet()) {
-
-      TextureInfo ti = new TextureInfo();
-      ti.textureUnit = glEngine.useTexture(this.gl4, textureInput.getValue());
-      ti.name = textureInput.getValue();
-      ti.channel = textureInput.getKey();
-      ti.uniformLocation =
-          this.gl4.glGetUniformLocation(this.shaderProgram.id, "iChannel" + ti.channel);
-
-      textures.add(ti);
-    }
   }
 
   // Run loop
@@ -460,7 +297,6 @@ public class GLShader {
     // hand the complete uniform list to OpenGL
     updateUniforms();
     render();
-    saveBackBuffer();
   }
 
   /** Activate this shader for rendering in the current context */
@@ -469,116 +305,25 @@ public class GLShader {
   }
 
   private void setUniforms() {
-    // Set audio waveform and fft data as a 512x2 texture on the specified audio
-    // channel if it's a shadertoy shader, or iChannel0 if it's a local shader.
-
-    // By Imperial Decree, the audio texture will heretofore always use texture unit
-    // GL_TEXTURE0, the model coordinate array will use GL_TEXTURE1 and the backbuffer
-    // texture will use GL_TEXTURE2. Other (shader-specific) textures will be automatically
-    // bound to sequential ids starting with GL_TEXTURE3.
-    //
-    // The audio texture can be used by all shaders, and stays bound to texture
-    // unit 0 throughout the Chromatik run. All we have to do to use it is add the uniform.
-    setUniform(UniformNames.AUDIO_CHANNEL, 0);
-
-    // use the current view's model coordinates texture which
-    // has already been loaded to the GPU by the texture cache manager.
-    // All we need to do is point at the right GL texture unit.
-    if (this.modelCoordsChanged) {
-      this.modelCoordsChanged = false;
-      setUniform(UniformNames.LX_MODEL_COORDS, this.modelCoordsTextureUnit);
-    }
-
-    // Update backbuffer texture data. This buffer contains the result of the
-    // previous render pass.  It is always bound to texture unit 2.
-    gl4.glActiveTexture(GL_TEXTURE2);
-    gl4.glEnable(GL_TEXTURE_2D);
-    gl4.glBindTexture(GL4.GL_TEXTURE_2D, backbufferHandle[0]);
-
-    gl4.glTexImage2D(
-        GL4.GL_TEXTURE_2D,
-        0,
-        GL4.GL_RGBA,
-        width,
-        height,
-        0,
-        GL4.GL_BGRA,
-        GL_UNSIGNED_BYTE,
-        backBuffer);
-
-    setUniform(UniformNames.BACK_BUFFER, 2);
-
-    // if necessary, update the mapped buffer texture data
-    if (useMappedBuffer) {
-      gl4.glActiveTexture(GL_TEXTURE0 + mappedBufferUnit);
-      gl4.glEnable(GL_TEXTURE_2D);
-      gl4.glBindTexture(GL4.GL_TEXTURE_2D, mappedBufferHandle[0]);
-
-      gl4.glTexImage2D(
-          GL4.GL_TEXTURE_2D,
-          0,
-          GL4.GL_RGBA,
-          mappedBufferWidth,
-          mappedBufferHeight,
-          0,
-          GL4.GL_BGRA,
-          GL_UNSIGNED_BYTE,
-          mappedBuffer);
-
-      setUniform(UniformNames.MAPPED_BUFFER, mappedBufferUnit);
-    }
-
-    // add shadertoy texture channels. These textures already statically bound to
-    // texture units so all we have to do is tell the shader which texture unit to use.
-    for (TextureInfo ti : textures) {
-      setUniform(UniformNames.CHANNEL + ti.channel, ti.textureUnit);
-      gl4.glUniform1i(ti.uniformLocation, ti.textureUnit);
-    }
-
-    // Add all preprocessed LX parameters from the shader code as uniforms
-    for (LXParameter customParameter : fragmentShader.parameters) {
-      setUniform(
-          customParameter.getLabel() + UniformNames.LX_PARAMETER_SUFFIX,
-          customParameter.getValuef());
-    }
-
-    // Set user uniforms last, giving user the option to override any default values
     for (UniformSource uniformSource : this.uniformSources) {
       uniformSource.setUniforms(this);
     }
   }
 
-  /** Set up geometry at frame generation time */
-  private void render() {
-    gl4.glBindVertexArray(vaohandles[0]);
-    gl4.glDrawElements(GL2.GL_TRIANGLES, INDICES.length, GL2.GL_UNSIGNED_INT, 0);
+  /** Pass all modified (and new) uniform values to OpenGL */
+  private void updateUniforms() {
+    for (Uniform uniform : this.uniforms) {
+      if (uniform.hasUpdate()) {
+        uniform.update();
+      }
+    }
   }
 
-  private void saveBackBuffer() {
-    backBuffer.rewind();
-    gl4.glReadBuffer(GL_BACK);
+  /** Child classes should run GL draw commands here */
+  protected abstract void render();
 
-    // using BGRA byte order lets us read int values from the buffer and pass them
-    // directly to LX as colors, without any additional work on the Java side.
-    gl4.glReadPixels(0, 0, width, height, GL_BGRA, GL_UNSIGNED_BYTE, backBuffer);
-  }
-
-  public ByteBuffer getBackBuffer() {
-    return backBuffer;
-  }
-
-  // Staging Uniforms: LX Model
-
-  /**
-   * Copy LXPoints' normalized coordinates into textures for use by shaders. Must be called by the
-   * parent pattern or effect at least once before the first frame is rendered. And should be called
-   * by the pattern's frametime run() function if the model has changed since the last frame.
-   *
-   * @param model Current LXModel of the calling context, which is a LXView or the global model
-   */
-  public void setModelCoordinates(LXModel model) {
-    this.modelCoordsTextureUnit = this.glEngine.getCoordinatesTexture(model);
-    this.modelCoordsChanged = true;
+  protected int getVaoHandle() {
+    return this.vaoHandles[0];
   }
 
   // Staging Uniforms
@@ -871,19 +616,6 @@ public class GLShader {
     return uniform;
   }
 
-  // Passing Uniforms to OpenGL
-
-  // parse uniform list and create necessary GL objects
-  // Note that array buffers passed in must be allocated to the exact appropriate size
-  // you want. No allocating a big buffer, then partly filling it. GL is picky about this.
-  private void updateUniforms() {
-    for (Uniform uniform : this.uniforms) {
-      if (uniform.hasUpdate()) {
-        uniform.update();
-      }
-    }
-  }
-
   // get a texture id for a uniform either from uniformTextureUnits or by allocating a new one
   // TODO: should these be unique per shader, instead of global unique?
   private int getTextureUnit(String name) {
@@ -896,33 +628,16 @@ public class GLShader {
     }
   }
 
-  // Releases native resources allocated by this shader.
-  // Should be called by the pattern's dispose() function
-  // when the pattern is unloaded. (Not when just
-  // deactivated.)
   public void dispose() {
     // Release references to uniform objects
     this.uniformMap.clear();
     this.mutableUniforms.clear();
 
-    // if we've been fully initialized, we need to release all
-    // OpenGL GPU resources we've allocated.
     if (this.initialized) {
-
       // delete GPU buffers we directly allocated
       gl4.glDeleteBuffers(2, geometryBufferHandles, 0);
-      gl4.glDeleteVertexArrays(1, vaohandles, 0);
-      gl4.glDeleteTextures(1, backbufferHandle, 0);
+      gl4.glDeleteVertexArrays(1, vaoHandles, 0);
 
-      if (useMappedBuffer) {
-        gl4.glDeleteTextures(1, mappedBufferHandle, 0);
-        glEngine.releaseTextureUnit(mappedBufferUnit);
-      }
-
-      // free any textures on ShaderToy channels
-      for (TextureInfo ti : textures) {
-        glEngine.releaseTexture(ti.name);
-      }
       this.shaderProgram.dispose();
     }
   }
