+++ te-app/src/main/java/titanicsend/pattern/glengine/GLShaderPattern.java
@@ -24,8 +24,8 @@ public class GLShaderPattern extends TEPerformancePattern {
   protected final int mappedBufferHeight = GLEngine.getMappedBufferHeight();
 
   // list of shaders to run
-  private final List<GLShader> mutableShaders = new ArrayList<>();
-  protected final List<GLShader> shaders = Collections.unmodifiableList(this.mutableShaders);
+  private final List<TEShader> mutableShaders = new ArrayList<>();
+  protected final List<TEShader> shaders = Collections.unmodifiableList(this.mutableShaders);
 
   private boolean modelChanged = true;
 
@@ -61,7 +61,7 @@ public class GLShaderPattern extends TEPerformancePattern {
     super(lx, view);
   }
 
-  private GLShader addShader(GLShader shader) {
+  private TEShader addShader(TEShader shader) {
     this.mutableShaders.add(shader);
     return shader;
   }
@@ -70,35 +70,35 @@ public class GLShaderPattern extends TEPerformancePattern {
    * Add a shader by fragment shader filename. The simple option for shaders that use only the
    * default TEPerformancePattern uniforms and don't require any additional computation in Java.
    */
-  protected GLShader addShader(String shaderName, String... textureFilenames) {
-    return addShader(new GLShader(lx, shaderName, this::setUniforms, textureFilenames));
+  protected TEShader addShader(String shaderName, String... textureFilenames) {
+    return addShader(new TEShader(lx, shaderName, this::setUniforms, textureFilenames));
   }
 
   /**
    * Add a shader by fragment shader filename. The simple option for shaders that use only the
    * default TEPerformancePattern uniforms and don't require any additional computation in Java.
    */
-  protected GLShader addShader(
-      String shaderName, GLShader.UniformSource uniformSource, String... textureFilenames) {
+  protected TEShader addShader(
+      String shaderName, TEShader.UniformSource uniformSource, String... textureFilenames) {
     return addShader(
-        new GLShader(lx, shaderName, List.of(this::setUniforms, uniformSource), textureFilenames));
+        new TEShader(lx, shaderName, List.of(this::setUniforms, uniformSource), textureFilenames));
   }
 
   /** Add a shader by fragment shader filename, with a callback for setting custom uniforms. */
-  protected GLShader addShader(String shaderName, GLShader.UniformSource uniformSource) {
-    return addShader(new GLShader(lx, shaderName, List.of(this::setUniforms, uniformSource)));
+  protected TEShader addShader(String shaderName, TEShader.UniformSource uniformSource) {
+    return addShader(new TEShader(lx, shaderName, List.of(this::setUniforms, uniformSource)));
   }
 
   /** Add a shader by fragment shader filename, with a callback for setting custom uniforms. */
-  protected GLShader addShader(
-      String shaderName, GLShader.UniformSource uniformSource, ByteBuffer frameBuf) {
+  protected TEShader addShader(
+      String shaderName, TEShader.UniformSource uniformSource, ByteBuffer frameBuf) {
     return addShader(
-        new GLShader(lx, shaderName, List.of(this::setUniforms, uniformSource), frameBuf));
+        new TEShader(lx, shaderName, List.of(this::setUniforms, uniformSource), frameBuf));
   }
 
   /** Add a shader by fragment shader filename */
-  protected GLShader addShader(String shaderName, ByteBuffer frameBuf) {
-    return addShader(new GLShader(lx, shaderName, this::setUniforms, frameBuf));
+  protected TEShader addShader(String shaderName, ByteBuffer frameBuf) {
+    return addShader(new TEShader(lx, shaderName, this::setUniforms, frameBuf));
   }
 
   @Override
@@ -114,14 +114,14 @@ public class GLShaderPattern extends TEPerformancePattern {
     // Update the model coords texture only when changed (and the first run)
     if (this.modelChanged) {
       this.modelChanged = false;
-      for (GLShader shader : this.shaders) {
+      for (TEShader shader : this.shaders) {
         shader.setModelCoordinates(m);
       }
     }
 
     // run the chain of shaders, except for the last one,
     // copying the output of each to the next shader's input texture
-    GLShader shader = null;
+    TEShader shader = null;
     for (int i = 0; i < this.shaders.size(); i++) {
       shader = this.shaders.get(i);
       shader.run();
@@ -205,7 +205,7 @@ public class GLShaderPattern extends TEPerformancePattern {
     // fix exception on slow startup
     if (this.colors == null) return;
     super.onActive();
-    for (GLShader shader : this.shaders) {
+    for (TEShader shader : this.shaders) {
       shader.onActive();
     }
   }
@@ -215,7 +215,7 @@ public class GLShaderPattern extends TEPerformancePattern {
     // fix exception on slow startup
     // TODO(jkb): remove this null check?
     if (this.colors == null) return;
-    for (GLShader shader : this.shaders) {
+    for (TEShader shader : this.shaders) {
       shader.onInactive();
     }
     super.onInactive();
@@ -223,7 +223,7 @@ public class GLShaderPattern extends TEPerformancePattern {
 
   @Override
   public void dispose() {
-    for (GLShader shader : this.shaders) {
+    for (TEShader shader : this.shaders) {
       shader.dispose();
     }
     super.dispose();
